moved commands package to the cli package - moved constants to internal package - created editable CLI defaults
