+++ cmd/sbot.go
@@ -1,9 +1,8 @@
 package cmd
 
 import (
+	"github.com/restechnica/semverbot/pkg/cli/commands"
 	"github.com/spf13/cobra"
-
-	"github.com/restechnica/semverbot/pkg/commands"
 )
 
 // NewApp creates a new semverbot CLI app
