+++ pkg/cli/commands/predict.go
