+++ pkg/cli/defaults.go
@@ -1,0 +1,14 @@
+package cli
+
+import "github.com/restechnica/semverbot/internal"
+
+var (
+	// DefaultConfig the default config.
+	DefaultConfig = internal.DefaultConfig
+
+	// DefaultConfigFilePath the default relative filepath to the config file.
+	DefaultConfigFilePath = internal.DefaultConfigFilePath
+
+	// DefaultVersion the default version when no other version can be found.
+	DefaultVersion = internal.DefaultVersion
+)
